[TTS][refactor] Part 3 - nemo.collections.tts.g2p.models (#6113)
* [TTS][refactor] Part 3 - nemo.collections.tts.g2p.models
* fixed invalid paths.
* update copyright header
---------
