+++ examples/tts/g2p/g2p_inference.py
@@ -21,7 +21,7 @@ import torch
 from omegaconf import OmegaConf
 from utils import get_metrics
 
-from nemo.collections.tts.g2p.models.g2p_model import G2PModel
+from nemo.collections.tts.models.base import G2PModel
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
