+++ examples/tts/g2p/g2p_train_and_evaluate.py
@@ -19,14 +19,14 @@ import torch
 from utils import get_model
 
 from nemo.collections.common.callbacks import LogEpochTimeCallback
-from nemo.collections.tts.g2p.models.g2p_model import G2PModel
+from nemo.collections.tts.models.base import G2PModel
 from nemo.core.config import hydra_runner
 from nemo.utils import logging, model_utils
 from nemo.utils.exp_manager import exp_manager
 
 """
 This script supports training of G2PModels 
-(for T5G2PModel use t5_g2p.yaml, for CTCG2PModel use either g2p_conformer.yaml or g2p_t5_ctc.yaml)
+(for T5G2PModel use g2p_t5.yaml, for CTCG2PModel use either g2p_conformer.yaml or g2p_t5_ctc.yaml)
 
 # Training T5G2PModel and evaluation at the end of training:
     python examples/text_processing/g2p/g2p_train_and_evaluate.py \
@@ -38,7 +38,7 @@ This script supports training of G2PModels
         do_training=True \
         do_testing=True
     
-    Example of the config file: NeMo/examples/text_processing/g2p/conf/t5_g2p.yaml
+    Example of the config file: NeMo/examples/tts/g2p/conf/g2p_t5.yaml
         
 # Training Conformer-G2P Model and evaluation at the end of training:
     python examples/text_processing/g2p/g2p_train_and_evaluate.py \
@@ -64,7 +64,7 @@ This script supports training of G2PModels
 """
 
 
-@hydra_runner(config_path="conf", config_name="t5_g2p")
+@hydra_runner(config_path="conf", config_name="g2p_t5")
 def main(cfg):
     trainer = pl.Trainer(**cfg.trainer)
     exp_manager(trainer, cfg.get("exp_manager", None))
