--- nemo/collections/tts/g2p/models/__init__.py
@@ -1,18 +1,0 @@
-# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from nemo.collections.tts.g2p.models.ctc_g2p import CTCG2PModel
-from nemo.collections.tts.g2p.models.t5_g2p import T5G2PModel
-
-__all__ = ["T5G2PModel", "CTCG2PModel"]
