+++ nemo/collections/tts/g2p/modules.py
@@ -79,7 +79,7 @@ class BaseG2p(ABC):
         """
 
         try:
-            from nemo.collections.tts.g2p.models.heteronym_classification import HeteronymClassificationModel
+            from nemo.collections.tts.models.g2p_heteronym_classification import HeteronymClassificationModel
 
             self.heteronym_model = heteronym_model
             self.heteronym_model.set_wordid_to_phonemes(wordid_to_phonemes_file)
